Add category and uploader feed
I made some changes to returne category and uploader feed. I also changed the URI to the magnet link to be able to use the feed in a torrent client. As discussed here (https://github.com/RSS-Bridge/rss-bridge/issues/412), I'd rather use the <torrent:magnetURI> item but it's not possible with RSS-Bridge ATM.

If I find time to work on it I'll try to add combination possibilities: search term in a certain category or for a specific uploader

Hope my changes will be appreciated!