Create KATBridge.php
Hi, I've created a bridge for KickassTorrents. It's mainly based on ThePirateBay bridge so a huge thanks to @mitsukarenai
This bridge have the same abilities: search terms (global or in category), categories, users, restrict results to trusted users.

I've tried to test it thoroughly and everything is working for me. I'm sure there is space for improvements and I'd gladly welcome any idea!

I have a concern about file size which are frequently reported on torrent site rss. I've added it to the description but it could be useful to add <size> or <length> tag.

Thanks again for RSSBridge, I'm loving it!