format: Return valid Media RSS instead of plain RSS 2.0 for 'Mrss'
The format MRSS did not return Media-RSS according to
http://www.rssboard.org/media-rss, but instead returned regular RSS
2.0 feeds. Feed readers expecting Media-RSS cannot work with that
format.

This adds RSS as a separate format via '&format=Rss' and changes the
existing MRSS format to adhere to the standards.

Notice: Existing feeds might break. If this is happening to you,
change the format from '&format=Mrss' to '&format=Rss'.