adding a Tumblr bridge
Why:
Tumblr has an RSS feature - always had -, and it's simple: username.tumblr.com/rss
However, when Verizon bought Yahoo, Oath placed an interstitial in front of most of Tumbl's content, including (!) the RSS feeds.

Unless... you're google. Or you look like Google.
This is a dead simple wrapper that has the 'Googlebot/' string in the User Agent to be able to fetch Tumblr RSS feeds.