[RadioMelodieBridge] Fix to use the new website layout
Two date issues fixed !

The regular eexpression did not match days with one digit (because the
day is always printed with two caracters, with a leading space), and did
not match months with accented characters like 'décembre'.

The date 'translate' function had a wrong placed duplicate for
'november' in the '$replace_en' array, which leaded to a wrong date conversion.