[ZoneTelechargementBridge] Fix SSL errors
The streaming unprotected URI had some SSL issue because the domain is
now hosting a parking website. This URL is the only known URL that allows to
show streaming links without being protected by Cloudflare DDoS protection.

This URL still works if you ask Cloudflare for the content. Therefore,
to load this URL, this bridge kindly ask cURL to connect to the same IP
as the unproteced DDL URI to get the Streaming URI content, by using the
CURLOPT_RESOLVE option.

So we are still able to get streamingg links ! (fix for #2419)