Add simple dependency injection container
This is just a very stupid DI container that creates requested objects
and passes them dependencies based on typehints of their constructor methods.

In the future, we may want to implement PSR-11 so that consumers of the library
can re-use their own dependency container instead of relying on this one.

https://www.php-fig.org/psr/psr-11/

Use singleton for now, until we can use DI everywhere.