Manual fixup of the rector-generated code
Rector tends to squash function call arguments to a single line but since there is a lot of arguments, it violates line length limit of our coding style. PHP CodeSniffer should be wise enough to split the arguments one-per-line but unfortunately it does not so let’s restore the original format manually.