[InstagramEmbedBridge] Add a new Instagram Bridge based on the embed url
This bridge is a simplified version of InstagramBridge that is based on
the embed page of a profile page.

It does only support the Username mode, and is limited to the last 6
medias.

At least actually, it does not need any cookies, and does work from a
server IP, without limitation.