[TldrTechBridge] Rename Webdev and add new channels
TLDR has renamed the 'Webdev' channel to 'Dev' and Webdev has stopped working in the bridge. This PR makes the applicable change.

They have also added new channels for 'Fintech' and 'Data' and 'Coming Soon' channels for 'IT' and 'Hardware'. The URLs and signup pages for the 'Coming Soon' channels exist but they haven't started publishing articles yet. This PR adds options for all of these new channels.