[GolemBridge] Fix multiple YouTube embeds
The following test article revealed that the previous code always takes
the first youtube link on all YouTube embed spaces:
https://www.golem.de/news/free-to-play-kostenlos-skaten-ballern-und-klicken-2510-201000.html

This command changes it to opportunistically select the i-th youtube
embed link for the i-th placeholder. This might fail and produce wrong
orderings, but should be better than the previous code. The above test
article is at least completely correct.