Send origin referrer for YouTube embeds
Unfortunately, YouTube blocks embedded views in iframes that don't send
a referrer. To fix them as-is, we explicitly provide the origin.

A future implementation might offer a config option to either offer a
YouTube video as iframe, requiring to send referrers towards Google, or
provide just a link (with thumbnail), which requires no referrer, iframe
or JavaScript.

https://support.google.com/youtube/answer/171780#zippy=%2Cprovide-an-http-referrer-header-to-enable-video-playback
https://developers.google.com/youtube/terms/required-minimum-functionality#embedded-player-api-client-identity