feat: optimize caching when TTL is zero
Currently setting the feed's cache TTL value to 0 will still save a copy of that feed to cache
with an expiration set to now. This is needless, since it will never get a chance to be
retrieved.

This change fixes that, together with some slight optimizations based on the assumption that
TTL=0 disables caching.

It also updates the docs on caching explaining this behavior.